pyproject.toml
[project]
name = "ryland"
version = "0.15.0"
description = "A simple static site generation library"
readme = "README.md"
license = "MIT"
authors = [
    { name = "James Tauber", email = "jtauber@jtauber.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.9",
    "pyyaml>=6.0.2",
    "Pillow>=12.0.0",
]

[project.urls]
github = "https://github.com/jtauber/ryland"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage>=7.10.6",
    "ruff>=0.12.12",
    "ty>=0.0.1a20",
]
